==> Instacart Basket Analysis/03 Scripts/1 IC Data Wrangling - customers.ipynb <==
# 1 IC Data Wrangling - customers

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Data Wrangling
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "customers" file

cust = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

# 03 First Look at Data

# Shape of "cust"

cust.shape

# First few rows of "cust"

cust.head()

# Data types of "cust"

cust.dtypes

# 04 Data Wrangling

### 01 Drop Columns

# Drop "First Name" and "Surnam" from "cust" for security reasons
# Drop "date_joined" from "cust" because it's irrelevant for this project

cust = cust.drop(columns =['First Name', 'Surnam', 'date_joined'])

cust.head()

### 02  Rename Columns

# Rename columns from "customers" to follow the same naming convention

# Create a variabel with new column names

column_names = {
    'Gender' : 'gender',
    'STATE' : 'state',
    'Age' : 'age'}

# Rename columns

cust.rename(columns = column_names, inplace=True)

cust.head()

### 03 Change Data Types

No data types are being changed for this dataframe.

# 05 Export Data

# Export the "cust" dataframe

cust.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'wrangled_customers.csv'))
