==> Instacart Basket Analysis/03 Scripts/1 IC Data Wrangling - departments.ipynb <==
# 1 IC Data Wrangling - departments

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Data Wrangling
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "deps" file

deps = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

# 03 First Look at Data

# Shape of "ords"

deps.shape

# First few rows of "ords"

deps.head()

# Data types of "ords"

deps.dtypes

# 04 Data Wrangling

### 01 Transpose Dataframe

# Transpose "deps" into a long format

deps = deps.T

deps

# Create a new index

deps.reset_index()

# Create a new header

new_header = deps.iloc[0]

# Keep the dataframe without the first row

deps = deps[1:]

deps

# Add the new header

deps.columns = new_header

deps

### 02 Drop Columns

No columns are being dropped in this dataframe.

### 03  Rename Columns

No columns are being renamed in this dataframe.

### 04 Change Data Types

No data types are being changed for this dataframe.

# 05 Export Data

# Export the "deps" dataframe

deps.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'wrangled_departments.csv'))
