==> Instacart Basket Analysis/03 Scripts/1 IC Data Wrangling - orders.ipynb <==
# 1 IC Data Wrangling - orders

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Data Wrangling
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "orders" file

ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# 03 First Look at Data

# Shape of "ords"

ords.shape

# First few rows of "ords"

ords.head()

# Data types of "ords"

ords.dtypes

# 04 Data Wrangling

### 01 Drop Columns

# Drop the "eval_set" column as it is irrelevant for this project

ords = ords.drop(columns = ['eval_set'])

ords.head()

### 02  Rename Columns

# Rename "orders_dow" to "orders_day_of_week" for better understanding

ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

ords.head()

### 03 Change Data Types

No data types are being changed for this dataframe.

# 05 Export Data

# Export the "ords" dataframe

ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'wrangled_orders.csv'))
