==> Instacart Basket Analysis/03 Scripts/1 IC Data Wrangling - orders_products_prior.ipynb <==
# 1 IC Data Wrangling - orders_products_prior

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Data Wrangling
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "orders_products_prior" file

ords_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'), index_col = False)

# 03 First Look at Data

# Shape of "ords_prods"

ords_prods.shape

# First few rows of "ords_prods"

ords_prods.head()

# Data types of "ords_prods"

ords_prods.dtypes

# 04 Data Wrangling

### 01 Drop Columns

No columns are being dropped for this dataframe.

### 02  Rename Columns

No columns are being renamed for this dataframe.

### 03 Change Data Types

No data type changes are being done for this dataframe.

# 05 Export Data

# Export the "ords_prods" dataframe

ords_prods.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'wrangled_orders_products.csv'))
