==> Instacart Basket Analysis/03 Scripts/2 IC Data Consistency Checks - orders.ipynb <==
# 2 IC Data Consistency Checks - orders

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Consistency Checks
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set the data path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "wrangled_orders" file

ords_wr = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'wrangled_orders.csv'), index_col = False)

# 03 First Look at Data

# Shape of "ords_wr"

ords_wr.shape

# First few rows of "ords_wr"

ords_wr.head()

# Drop "Unnamed: 0" column

ords_wr = ords_wr.drop(columns =['Unnamed: 0'])

ords_wr.head()

# Data types of "ords_wr"

ords_wr.dtypes

# 04 Consistency Checks

# Descriptive statistics of "ords_wr"

ords_wr.describe()

No obvious problems were detected.

### 01 Mixed-Type Data

# Check for mixed-type data

for col in ords_wr.columns.tolist():
    weird = (ords_wr[[col]].map(type) != ords_wr[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (ords_wr[weird]) > 0:
        print (col)

No mixed-type data could be found.

### 02 Missing Values

# Check for missing values in "ords_wr"

ords_wr.isnull().sum()

There are 206.209 missing values in "days_since_prior_order". Since the first order of a customer doesn't have a value for this column, as there is no prior order, there would be a missing value for that case.
To check if this is the case, a crosstab will be created.

# Create a crosstab for "days_since_prior_order" and "order_number"

crosstab = pd.crosstab(ords_wr['days_since_prior_order'], ords_wr['order_number'], dropna = False)

# Copy the table to clipboard

crosstab.to_clipboard()

The crosstab proved that the missing values in the "days_since_prior_order" column are due to the fact that there will be no prior order before day one. A screenshot will be included in the final report.

Since the rows with the missing values still contain valuable information, they will be left in the data set.

### 03 Duplicate Data

# Create a new subset with only duplicates

ords_wr_dups = ords_wr[ords_wr.duplicated()]

ords_wr_dups

There is no duplicate data in this data set.

# Number of rows after consistency check

ords_wr.shape

# 05 Export Data

# Export "ords_wr"

ords_wr.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'checked_orders.csv'))
