==> Instacart Basket Analysis/03 Scripts/2 IC Data Consistency Checks - orders_products_prior.ipynb <==
# 2 IC Data Consistency Checks - orders_products_prior

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Consistency Checks
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set the data path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "wrangled_orders_products" file

ords_prods_wr = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'wrangled_orders_products.csv'), index_col = False)

# 03 First Look at Data

# Shape of "ords_prods_wr"

ords_prods_wr.shape

# First few rows of "ords_prods_wr"

ords_prods_wr.head()

# Drop "Unnamed: 0" column

ords_prods_wr = ords_prods_wr.drop(columns =['Unnamed: 0'])

ords_prods_wr.head()

# Data types of "ords_prods_wr"

ords_prods_wr.dtypes

# 04 Consistency Checks

# Descriptive statistics of "ords_prods_wr"

ords_prods_wr.describe()

No obvious problems were detected.

### 01 Mixed-Type Data

# Check for mixed-type data

for col in ords_prods_wr.columns.tolist():
    weird = (ords_prods_wr[[col]].map(type) != ords_prods_wr[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (ords_prods_wr[weird]) > 0:
        print (col)

No mixed-type data could be found.

### 02 Missing Values

# Check for missing values in "ords_prods_wr"

ords_prods_wr.isnull().sum()

No missing values could be found.

### 03 Duplicate Data

# Create a new subset with only duplicates

ords_prods_wr_dups = ords_prods_wr[ords_prods_wr.duplicated()]

ords_prods_wr_dups

There is no duplicate data in this data set.

# Number of rows after consistency check

ords_prods_wr.shape

# 05 Export Data

# Export "ords_prods_wr"

ords_prods_wr.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'checked_orders_products.csv'))
