==> Instacart Basket Analysis/03 Scripts/2 IC Data Consistency Checks - products.ipynb <==
# 2 IC Data Consistency Checks - products

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Consistency Checks
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set the data path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "wrangled_products" file

prods_wr = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'wrangled_products.csv'), index_col = False)

# 03 First Look at Data

# Shape of "prods_wr"

prods_wr.shape

# First few rows of "prods_wr"

prods_wr.head()

# Drop "Unnamed: 0" column

prods_wr = prods_wr.drop(columns =['Unnamed: 0'])

prods_wr.head()

# Data types of "prods_wr"

prods_wr.dtypes

# 04 Consistency Checks

# Descriptive statistics of "prods_wr"

prods_wr.describe()

The maximum value for "prices" is 99.999$ which is clearly too high.

##### Fix "prices" values

# Check for outliers (prices > $100)

prods_wr.loc[prods_wr['prices'] > 100]

# Mark outliers as missing

prods_wr.loc[prods_wr['prices'] > 100, 'prices'] = np.nan

# Check for success by looking for max price value

prods_wr['prices'].max()

prods_wr['prices'].describe()

### 01 Mixed-Type Data

# Check for mixed-type data

for col in prods_wr.columns.tolist():
    weird = (prods_wr[[col]].map(type) != prods_wr[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (prods_wr[weird]) > 0:
        print (col)

prods_wr['product_name'].dtype

The column "product_name" seems to have mixed-type data. When checking for the data type, the results shows that this column has the data type "string".

### 02 Missing Values

# Check for missing values in "prods_wr"

prods_wr.isnull().sum()

There are 16 missing values in the "product_name" column and 2 missing values in the "prices" column. The latter is due to the fact that I marked the outlier values as missing.

Since the column "product_name" has missing data, it is possible that that is the reason why the column showed up as mixed-type data.

# Create a subset with only missing values in "prods_wr"

nan = prods_wr[prods_wr['product_name'].isnull() == True]

nan

# Create a subset of the data with no missing values

# Create a subset with no missing values

prods_wr_clean = prods_wr[prods_wr['product_name'].isnull() == False]

# Check again for mixed-type data

for col in prods_wr_clean.columns.tolist():
    weird = (prods_wr_clean[[col]].map(type) != prods_wr_clean[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (prods_wr_clean[weird]) > 0:
        print (col)

After creating a new dataframe without the rows with missing data, there is no more mixed-type data.

# Check the row number
prods_wr_clean.shape

This new dataframe has 16 rows less, the exact amount of rows containing missing values in the "prices" column.

### 03 Duplicate Data

# Create a new subset with only duplicates

prods_wr_dups = prods_wr[prods_wr.duplicated()]

prods_wr_dups

# Create a dataframe without duplicates

prods_wr_no_dups = prods_wr_clean.drop_duplicates()

# Number of rows after consistency check

prods_wr_no_dups.shape

# 05 Export Data

# Export "prods_wr"

prods_wr_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'checked_products.csv'))
