==> Instacart Basket Analysis/03 Scripts/3 IC Merging - merged_orders_products_orders with products.ipynb <==
# 3 IC Merging - merged_orders_products_orders with products

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data and Downsizing of Data Types
#### 04 Merge
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "checked_products" file

prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'checked_products.csv'), index_col = False)

# Import the "merged_orders_products_orders" file

ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_orders_products_orders.pkl'))

# 03 First Look at Data and Downsizing of Data Types

##### "prods"

# Shape of "prods"

prods.shape

# First few rows of "prods"

prods.head()

# Drop "Unnamed: 0" column

prods = prods.drop(columns =['Unnamed: 0'])

prods.head()

# Data types of "prods"

prods.dtypes

# Dowsize data types for a better merging

prods['product_id'] = prods['product_id'].astype('int32')
prods['aisle_id'] = prods['aisle_id'].astype('int8')
prods['department_id'] = prods['department_id'].astype('int8')

# Check the data types

prods.dtypes

##### "ords_prods"

# Shape of "ords_prods"

ords_prods.shape

# First few rows of "ords_prods"

ords_prods.head()

# Drop "_merge" column

ords_prods = ords_prods.drop(columns =['_merge'])

ords_prods.head()

# Data types of "ords_prods"

ords_prods.dtypes

# 04 Merge

# Merge the two dataframes

merged = prods.merge(ords_prods, on = 'product_id', indicator = True)

# Have a look at the new dataframe

merged.shape

merged.head()

# 05 Export Data

# Export "merged"

merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_orders_products_products.pkl'))
