==> Instacart Basket Analysis/03 Scripts/3 IC Merging - merged_orders_products_products with customers.ipynb <==
# 3 IC Merging - orders with orders_products_prior

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data and Downsizing of Data Types
#### 04 Merge
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "checked_customers" file

cust = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'checked_customers.csv'), index_col = False)

# Import the "merged_orders_products_products" file

ords_prods_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_orders_products_products.pkl'))

# 03 First Look at Data and Downsizing of Data Types

##### "ords"

# Shape of "cust"

cust.shape

# First few rows of "cust"

cust.head()

# Drop "Unnamed: 0" column

cust = cust.drop(columns =['Unnamed: 0'])

cust.head()

# Data types of "cust"

cust.dtypes

# Dowsize data types for a better merging

cust['user_id'] = cust['user_id'].astype('int32')
cust['age'] = cust['age'].astype('int8')
cust['n_dependants'] = cust['n_dependants'].astype('int8')

# Check the data types

cust.dtypes

# Check for irregularities due to downsizing

cust.describe()

##### "ords_prods"

# Shape of "ords_prods_prods"

ords_prods_prods.shape

# First few rows of "ords_prods_prods"

ords_prods_prods.head()

# Data types of "ords_prods_prods"

ords_prods_prods.dtypes

# Drop "_merge" column

ords_prods_prods = ords_prods_prods.drop(columns =['_merge'])

ords_prods_prods.head()

# 04 Merge

# Merge the two dataframes

merged = cust.merge(ords_prods_prods, on = 'user_id', indicator = True)

# Have a look at the new dataframe

merged.shape

merged.head()

# 05 Export Data

merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_all.pkl'))
