==> Instacart Basket Analysis/03 Scripts/3 IC Merging - orders with orders_products_prior.ipynb <==
# 3 IC Merging - orders with orders_products_prior

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data and Downsizing of Data Types
#### 04 Merge
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import the "checked_orders" file

ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'checked_orders.csv'), index_col = False)

# Import the "checked_orders_products" file

ords_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'checked_orders_products.csv'), index_col = False)

# 03 First Look at Data and Downsizing of Data Types

##### "ords"

# Shape of "ords"

ords.shape

# First few rows of "ords"

ords.head()

# Drop "Unnamed: 0" column

ords = ords.drop(columns =['Unnamed: 0'])

ords.head()

# Data types of "ords"

ords.dtypes

# Dowsize data types for a better merging

ords['order_id'] = ords['order_id'].astype('int32')
ords['user_id'] = ords['user_id'].astype('int32')
ords['order_number'] = ords['order_number'].astype('int8')
ords['orders_day_of_week'] = ords['orders_day_of_week'].astype('int8')
ords['order_hour_of_day'] = ords['order_hour_of_day'].astype('int8')

# Check the data types

ords.dtypes

##### "ords_prods"

# Shape of "ords_prods"

ords_prods.shape

# First few rows of "ords_prods"

ords_prods.head()

# Drop "Unnamed: 0" column

ords_prods = ords_prods.drop(columns =['Unnamed: 0'])

ords_prods.head()

# Data types of "ords_prods"

ords_prods.dtypes

# Dowsize data types for a better merging

ords_prods['order_id'] = ords_prods['order_id'].astype('int32')
ords_prods['product_id'] = ords_prods['product_id'].astype('int32')
ords_prods['add_to_cart_order'] = ords_prods['add_to_cart_order'].astype('int32')
ords_prods['reordered'] = ords_prods['reordered'].astype('int8')

# Check the data types

ords_prods.dtypes

# 04 Merge

# Merge the two dataframes

merged = ords.merge(ords_prods, on = 'order_id', indicator = True)

# Have a look at the new dataframe

merged.shape

merged.head()

# 05 Export Data

merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_orders_products_orders.pkl'))
