==> Instacart Basket Analysis/03 Scripts/4 IC Analysis - creating flags.ipynb <==
# 4 IC Analysis - creating flags

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Create Flags
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import "merged_all"

merged_all = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_all.pkl'))

# 03 First Look at Data

# Shape of "merged_all"

merged_all.shape

# First few rows of "merged_all"

merged_all.head()

# Drop "_merge" column

merged_all = merged_all.drop(columns = ['_merge'])

merged_all.head()

# Data types of "merged_all"

merged_all.dtypes

# 04 Create Flags

##### price_range_flag

# Create conditions for a "price_range" flag

merged_all.loc[merged_all['prices'] > 15, 'price_range_flag'] = 'High range product'
merged_all.loc[(merged_all['prices'] <= 15) & (merged_all['prices'] > 5), 'price_range_flag'] = 'Mid range product'
merged_all.loc[merged_all['prices'] <= 5, 'price_range_flag'] = 'Low range product'

# Check the price range frequency

merged_all['price_range_flag'].value_counts(dropna = False)

# Check the nans

nan = merged_all[merged_all['prices'].isna()]

nan['product_name'].value_counts(dropna = False)

There seem to be two products that don't have a price. They are probably the ones that I changed from too high values to NaN.

##### busiest_day_flag

# Check order number by day

merged_all['orders_day_of_week'].value_counts(dropna = False)

# Create a for-loop for busyness of day

result = []

for value in merged_all["orders_day_of_week"]:
    if value == 0:
        result.append("Busiest day")
    elif value == 4:
        result.append("Least busy")
    else:
        result.append("Regularly busy")

# Print result list

result

# Combine result list with "merged_all"

merged_all['busiest_day_flag'] = result

# Check the frequency of busyness

merged_all['busiest_day_flag'].value_counts(dropna = False)

##### busiest_period_flag

# Check order number by hour

merged_all['order_hour_of_day'].value_counts(dropna = False)

##### Most Orders: 10h - 16h
##### Fewest Orders: 2h - 5h

# Create a for-loop for busyness by hour

result2 = []

for value in merged_all["order_hour_of_day"]:
    if value in [10, 11, 12, 13, 14, 15, 16]:
        result2.append("Most orders")
    elif value in [2, 3, 4, 5]:
        result2.append("Fewest orders")
    else:
        result2.append("Average orders")

# Print result2

result2

# Combine result2 with "merged_all"

merged_all['busiest_period_flag'] = result2

# Check the frequency of busyness

merged_all['busiest_period_flag'].value_counts(dropna = False)

##### loyalty_flag

# Group "merged_all" by "user_id" and create a column with the max order numbers

merged_all['max_order'] = merged_all.groupby(['user_id'])['order_number'].transform(np.max)

merged_all.head()

# Create a flag that assigns a "loyalty label" to a user ID

merged_all.loc[merged_all['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'
merged_all.loc[(merged_all['max_order'] <= 40) & (merged_all['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'
merged_all.loc[merged_all['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# Check the loyalty frequency

merged_all['loyalty_flag'].value_counts(dropna = False)

##### spending_flag

# Group "merged_all" by "user_id" and create a column with the average spending

merged_all['avg_spending'] = merged_all.groupby(['user_id'])['prices'].transform(np.mean)

merged_all.head()

# Create a flag that assigns a "spending label" to a user ID

merged_all.loc[merged_all['avg_spending'] < 10, 'spending_flag'] = 'Low spender'
merged_all.loc[merged_all['avg_spending'] >= 10, 'spending_flag'] = 'High spender'

# Check the spender frequency

merged_all['spending_flag'].value_counts(dropna = False)

##### order_frequency_flag

# Group "merged_all" by "user_id" and create a column with the median of days since prior order

merged_all['median_days_since_prior_order'] = merged_all.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

merged_all.head()

# Create a flag that assigns a "frequency label" to a user ID

merged_all.loc[merged_all['median_days_since_prior_order'] > 20, 'frequency_flag'] = 'Non-frequent customer'
merged_all.loc[(merged_all['median_days_since_prior_order'] > 10) & (merged_all['median_days_since_prior_order'] <= 20), 'frequency_flag'] = 'Regular customer'
merged_all.loc[merged_all['median_days_since_prior_order'] <= 10, 'frequency_flag'] = 'Frequent customer'

# Check the frequency of order frequency

merged_all['frequency_flag'].value_counts(dropna = False)

There are 5 rows that couldn't be assigned a flag.

# Find the rows with the NaN in the "frequency_flag" column

merged_all[merged_all['median_days_since_prior_order'].isna()]

This is one customer who ordered 5 different products in one order. Therefore, the only value for "days_since_prior_order" there is for this customer is "NaN". And since there can't be a median of NaN, this customer didn't get a frequency flag.

# Drop "max_order", "avg_spending" and "median_days_since_prior_order" to save space in the memory

merged_all = merged_all.drop(columns = ['max_order', 'avg_spending', 'median_days_since_prior_order'])

merged_all.head()

# 05 Export Data

# Export the dataframe

merged_all.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_all_flags.pkl'))
