==> Instacart Basket Analysis/03 Scripts/5 Tidy up the dataframe.ipynb <==
# 5 Tidy up the dataframe

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 First Look at Data
#### 04 Drop Columns
#### 05 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'D:\Data'

# Import "ic_keep"

ic_keep = pd.read_pickle(os.path.join(path, 'ic_keep.pkl'))

# First Look at Data

# Row names of "ic_keep"

ic_keep.dtypes

# Drop Columns

# Drop unnecessary columns

ic_keep = ic_keep.drop(columns = ['user_id', 'gender', 'product_id', 'product_name', 'aisle_id', 'add_to_cart_order', 'reordered', 'price_range_flag', 'busiest_day_flag', 'busiest_period_flag', 'exclusion_flag'])

# Check remaining columns of "ic_keep"

ic_keep.dtypes

# Export Data

# Export "ic_keep"

ic_keep.to_pickle(os.path.join(path, 'ic_tidy.pkl'))
