==> Instacart Basket Analysis/03 Scripts/6 Create a Subset to investigate.ipynb <==
# 6 Create a subset to investigate

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 Create a Subset
#### 04 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'D:\Data'

# Import "ic_keep"

ic_keep = pd.read_pickle(os.path.join(path, 'ic_keep.pkl'))

# 03 Create a subset

# Create a list holding True/False values to the test np.random.rand() <= 0.7

dev = np.random.rand(len(ic_keep)) <= 0.7

# Store 30% of the sample in the dataframe small

ic_small = ic_keep[~dev]

# Delete not needed dataframe to save memory

del ic_keep

# 04 Export Data

# Export "ic_small"

ic_small.to_pickle(os.path.join(path, 'ic_small.pkl'))
