==> Instacart Basket Analysis/03 Scripts/8 Visualization Task 4.10 Step 4.1.ipynb <==
# 8 Visualization Task 4.10 Step 4.1

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 Analysis
#### 04 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import "ic"

ic = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ic.pkl'))

# 03 Analysis

### 4.Create an exclusion flag for low-activity customers (customers with less than 5 orders) and exclude them from the data.

# Create a flag that assigns an "exclusion label" to a user ID

ic.loc[ic['max_order'] < 5, 'exclusion_flag'] = 'exclude'
ic.loc[ic['max_order'] >= 5, 'exclusion_flag'] = 'keep'

# Check the exclusion frequency

ic['exclusion_flag'].value_counts(dropna = False)

# 04 Export Data

# Export the "ic_4" dataframe

ic.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ic_4.pkl'))
