==> Instacart Basket Analysis/03 Scripts/8 Visualization Task 4.10 Step 4.2.ipynb <==
# 8 Visualization Task 4.10 Step 4.2

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 Analysis
#### 04 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import "ic_4"

ic = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ic_4.pkl'))

# 03 Analysis

# Create a dataframe with only rows that are marked as "keep" in the "exclusion_flag"

ic_keep = ic[ic['exclusion_flag'] == 'keep']

# Check the exclusion frequency for a successful exclusion

ic_keep['exclusion_flag'].value_counts(dropna = False)

# Delete the big dataframe to save memory

del ic

# 04 Export Data

# Export the "ic_keep" dataframe

ic_keep.to_pickle(r'D:\Data\ic_keep.pkl')
