==> Instacart Basket Analysis/03 Scripts/8 Visualization Task 4.10 Steps 1-3.ipynb <==
# 8 Visualization Task 4.10 Steps 1-3

### This script contains the following points
#### 01 Import Libraries
#### 02 Import Data
#### 03 Analysis
#### 04 Export Data

# 01 Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

# 02 Import Data

# Set a path

path = r'C:\Users\Tina\Desktop\CareerFoundry\Data Analytics Immersion\Instacart Basket Analysis'

# Import "merged_all_flags"

ic = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'merged_all_flags_short.pkl'))

# 03 Analysis

### 2. Consider any security implications that might exist for this new data. You'll need to address any PII data in the data before continuing your anlysis.

Any PII data from this data set has already been dropped.

### 3. Is there a difference in spending habits between the different U.S. regions?

# Create a "region" column that divides the "state" column into geographical regions


result = []

for value in ic ['state']:
    if value in ['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey']:
        result.append('northeast')
    elif value in ['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri']:
        result.append('midwest')
    elif value in ['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana']:
        result.append('south')
    elif value in ['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 'California', 'Hawaii']:
        result.append('west')
    else:
        result.append('unknown')

# Print result list

result

# Combine result list with "ic"

ic['region'] = result

# Check the frequency of regions

ic['region'].value_counts(dropna = False)

# Create a crosstable between "region" and "spending_flag"

region_spending = pd.crosstab(ic['region'], ic['spending_flag'], dropna = False)

region_spending

All the regions have more low spenders than they have high spenders. The region with the most high spenders is the south, the region with the least high spenders is the northeast. This is most likely due to the size of those regions which results in different sizes in population and therefore possible customers.

# 04 Export Data

# Export the "ic_keep" dataframe

ic.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ic.pkl'))
